This api manages all operations with error handling and modifies auto increment for all table IDs

+++ env/src/endpoints/events.py
@@ -1,5 +1,5 @@
-from fastapi import APIRouter
-from src.models.events_model import Event_model, conn
+from fastapi import APIRouter, HTTPException
+from src.models.events_model import Event_model,Event_model2, conn
 from typing import Optional
 
 router = APIRouter(
@@ -20,35 +20,27 @@ def create(event: Event_model):
     return event
 
 
-@router.put("/{id}", response_model=Event_model)
-def update_events(id: int, event: Event_model):
-    cursor = conn.cursor()
-    query = "UPDATE events SET id = %s, user_id = %s, name = %s, date = %s, location = %s WHERE id = %s"
-    cursor.execute(query, (event.id, event.user_id, event.name, event.date, event.location, id))
-    conn.commit()
-    cursor.close()
-    event.id = id
-
-    return event
-
-
-@router.get("/{id}", response_model=Event_model)
+@router.get("/{id}", response_model=Event_model2)
 def read_one(id: int):
     cursor = conn.cursor()
     query = "SELECT id, user_id, name, date, location FROM events WHERE id = %s"
     cursor.execute(query, (id,))
     event = cursor.fetchone()
+    print(event)
     cursor.close()
+    if event is None:
+        raise HTTPException(status_code=404, detail="Event not found")
     
     return {"id": event[0], "user_id": event[1], "name": event[2], "date": event[3], "location": event[4]}
 
 
-@router.get("/", response_model=list[Event_model])
+@router.get("/", response_model=list[Event_model2])
 def read_all():
     cursor = conn.cursor()
     query = "SELECT * FROM events"
     cursor.execute(query)
     events = cursor.fetchall()
+    print(events)
     events_data = []
     for event in events:
         event_dict = {
@@ -63,28 +55,58 @@ def read_all():
 
     return events_data
 
-
-@router.delete("/{id}")
-def delete_event(id: int):
+@router.put("/{id}", response_model=Event_model)
+def update_events(id: int, event: Event_model):
     cursor = conn.cursor()
-    query = "DELETE FROM events WHERE id = %s"
+    query = 'SELECT * FROM events WHERE id = %s';
     cursor.execute(query, (id,))
-    conn.commit()
-    cursor.close()
-
-    return {"id": id}
-
-
+    check_id = cursor.fetchone()
+    if check_id != None:
+        query = "UPDATE events SET id = id, user_id = %s, name = %s, date = %s, location = %s WHERE id = %s"
+        cursor.execute(query, (event.user_id, event.name, event.date, event.location, id))
+        conn.commit()
+        cursor.close()
+        event.id = id
+        return event
+    else:
+        raise HTTPException(status_code=404, detail="Event not found")
 
 @router.get("/report/{user_id}")
 def event_report(user_id: int):
     cursor = conn.cursor()
-    query = f"SELECT COUNT(id) FROM events WHERE user_id = %s;"
+    query = 'SELECT * FROM events WHERE user_id = %s';
     cursor.execute(query, (user_id,))
-    count = cursor.fetchone()[0]
-    conn.commit()
-    cursor.close()
+    check_id = cursor.fetchone()
+    if check_id != None:
+        query = f"SELECT COUNT(id) FROM events WHERE user_id = %s;"
+        cursor.execute(query, (user_id,))
+        count = cursor.fetchone()[0]
+        conn.commit()
+        cursor.close()
+        return {"count": count}
+    else:
+        raise HTTPException(status_code=404, detail="User  not found")
 
-    return {"count": count}
+
+
+@router.delete("/{id}")
+def delete_event(id: int):
+    cursor = conn.cursor()
+    event_id = id
+    query2 = 'SELECT * FROM events WHERE id = %s';
+    cursor.execute(query2, (id,))
+    check_id = cursor.fetchone()
+    if check_id != None:
+        query1 = 'DELETE FROM contributions WHERE event_id = %s';
+        cursor.execute(query1, (event_id,))
+        query = "DELETE FROM events WHERE id = %s"
+        cursor.execute(query, (id,))
+        conn.commit()
+        cursor.close()
+        return {"id": id}
+    else:
+        raise HTTPException(status_code=404, detail="Event not found")
+
+
